+++ CHANGELOG.md
@@ -1,6 +1,30 @@
 # Change Log
 
-## [0.6.6]
+## [0.6.8] - 17 Feb 2020
+
+- Oh my god !!! Did I just forget to break a switch statement ? _smh_.
+- Remember to run context again on generated function body to add code to source file guys <3.
+- Sidenote : Always make sure you exit switch statements !!!
+
+## [0.6.7] - 17 Feb 2020
+
+---
+
+- Removed complicated settings. Edits made to the **.code-workspace** file is now fully customizable by you using `SF.excludeFolders`, `SF.searchExclude`, `SF.excludedExtensions`.
+- This means, unless you are working on Editor extensions, you get best results using the default settings itself.
+- Patched a bug where Context won't activate with components with default values in header
+
+### Feature Updates:
+
+- Added `UE4 : Add Override function`. Helps adding { OnConstruction, Tick, BeginPlay etc.}
+- Also **supports context key**. Activates from lines with { public: , private: , protected: } in header.
+- Can be extended with `Functions_Ext.json` in sidebar. See `Functions_Core.json`(ReadOnly) for default list.
+- Added `SF.excludeFolders`, `SF.searchExclude`, `SF.excludedExtensions` for more control.
+- Removed `SF.allowUserExtensions`. Modding support is always on from now.
+
+## [0.6.6] - 16 Feb 2020
+
+---
 
 - Modders can now add extensions right from the editor sidebar.
 - Folders for AssetStreams are automatically added to sidebar under UE4.